+++ apps/admin-x-framework/src/index.ts
@@ -5,7 +5,7 @@ export {FrameworkProvider, useFramework} from './providers/FrameworkProvider';
 // Routing
 export type {RouteObject} from 'react-router';
 export type {RouterProviderProps} from './providers/RouterProvider';
-export {RouterProvider, useNavigate} from './providers/RouterProvider';
+export {RouterProvider, useNavigate, useBaseRoute, useIsBaseRoute} from './providers/RouterProvider';
 export {useNavigationStack} from './providers/NavigationStackProvider';
 export {Link, NavLink, Navigate, Outlet, useLocation, useParams, useSearchParams, redirect, matchRoutes} from 'react-router';
